+++ cloud-hypervisor/tests/common/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2025 The Cloud Hypervisor Authors. All rights reserved.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) mod tests_wrappers;
+pub(crate) mod utils;
